refactor: extract shared plugin utilities to reduce duplication
